move logs context to common
