+++ pkg/security/common/logs_context.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-package command
+package common
 
 import (
 	"fmt"
